Split project into modules
* `mod data` contains data structures
* `mod agg` contains aggregation functions
